+++ HPSU/plugins/pdf.py
